Avoid teardown of deleted channel

+++ monitoring/api/v3/alerts-client/snippets_test.py
@@ -63,8 +63,9 @@ class PochanFixture:
     def __exit__(self, type, value, traceback):
         # Delete the policy and channel we created.
         self.alert_policy_client.delete_alert_policy(self.alert_policy.name)
-        self.notification_channel_client.delete_notification_channel(
-            self.notification_channel.name)
+        if self.notification_channel.name:
+            self.notification_channel_client.delete_notification_channel(
+                self.notification_channel.name)
 
 
 @pytest.fixture(scope='session')
@@ -122,3 +123,4 @@ def test_delete_channels(capsys, pochan):
         pochan.project_name, [notification_channel_id], force=True)
     out, _ = capsys.readouterr()
     assert "{0} deleted".format(notification_channel_id) in out
+    self.notification_channel.name = None  # So teardown is not tried
